// src/app/client/appointment/rx-appointment-form.ts
import { FormControl } from "@angular/forms";

export interface RxAppointmentForm {
  clinic: FormControl<string>
  patient: FormControl<string>
  dentist: FormControl<string>
  services: FormControl<string[]>
  date: FormControl<Date>
  time: FormControl<string>
  patientNotes: FormControl<string>;
}